web/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.1.2
Flask-JWT-Extended==4.4.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.10
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
